+++ food-donator-backend/src/main/java/com/fooddonator/restapi/model/Donation.java
@@ -3,13 +3,13 @@ package com.fooddonator.restapi.model;
 public class Donation {
   public String id;
   public String userid;
-  public Integer donationdate;
+  public Long donationdate;
   public String starttime;
   public String endtime;
 
   public Donation() {}
 
-  public Donation(String _id, String _userid, Integer _donationDate, String _starttime, String _endtime) {
+  public Donation(String _id, String _userid, Long _donationDate, String _starttime, String _endtime) {
     id = _id;
     userid = _userid;
     donationdate = _donationDate;
